importer_app/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
requests==2.26.0
psycopg2-binary==2.9.1
Werkzeug==2.2.2
pandas
